requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
motor>=3.3.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
pytz>=2023.3
